Add MTF Autostretch feature to DeepSkyStacker
This commit adds a complete MTF (Midtone Transfer Function) stretch implementation
to DeepSkyStacker, providing a powerful alternative to the existing Asinh stretch.

Key features:
- MTF stretch tab with independent RGB channel control via gradient sliders
- One-click autostretch with configurable target background and shadow clipping
- Link/unlink button for synchronized channel adjustment
- Real-time preview with debounced slider updates
- Support for both manual MTF stretch and automatic parameter calculation

Implementation details:
- New mtfstretch.cpp algorithm in DeepSkyStackerKernel implementing:
  * MTF stretch formula: ((mtf - 1) * x) / ((2 * mtf - 1) * x - mtf)
  * Automatic parameter calculation using median and MAD statistics
  * RGB and monochrome image support with optional channel linking
- Custom QMTFSlider widget for three-handle gradient control (shadows/midtones/highlights)
- MtfAutostretchSettings dialog for configuring autostretch parameters
- Integration with existing ProcessingDlg preview and undo/redo system
- Proper signal debouncing using QTimer to avoid excessive preview updates

UI additions:
- MTF Stretch tab in ProcessingControls.ui with RGB gradient sliders
- Spinboxes for precise numeric control of shadows/midtones/highlights
- Link button (chain icon) to synchronize RGB channels
- Autostretch button with context menu for settings
- Clipping percentage labels showing shadow/highlight clipping
- Apply button to commit changes to undo/redo stack

Bug fixes included:
- asinhBeta/asinhBP now correctly initialize to DefaultAsinhBeta/DefaultAsinhBP
- zeroAsinHControls() resets to defaults instead of zero
- cbApplyPressed() properly disables button and uses emit
- Consistent Allman brace style throughout MTF code
- Added QMTFSlider to CMakeLists.txt Tools section

Files added:
- DeepSkyStacker/QMTFSlider.{h,cpp} - Custom gradient slider widget
- DeepSkyStacker/MtfAutostretchSettings.{h,cpp,ui} - Autostretch config dialog
- DeepSkyStacker/Icon/chain{,-linked}.svg - Link button icons
- DeepSkyStackerKernel/mtfstretch.cpp - Core MTF algorithm

Files modified:
- DeepSkyStacker/ProcessingDlg.{h,cpp} - Integration with MTF UI and controls
- DeepSkyStacker/ui/ProcessingControls.ui - New MTF stretch tab
- DeepSkyStacker/{CMakeLists.txt,*.vcxproj*} - Build system registration
- DeepSkyStackerKernel/{StackedBitmap.h,CMakeLists.txt} - MTF method declarations
- DeepSkyStackerTest/{AvxEntropyTest,PixelIteratorTest}.cpp - Type conversion fixes

+++ DeepSkyStackerTest/AvxEntropyTest.cpp
@@ -26,7 +26,7 @@ float roundSig(const float value, const float mult) {
 }
 
 float truncSig(const float value, const float mult) {
-	return static_cast<int>(value * mult);
+	return static_cast<float>(static_cast<int>(value * mult));
 }
 
 TEST_CASE("AVX Entropy", "[AVX][Entropy]")
